requirements.txt
azure-cosmos~=4.3.0
pymssql~=2.2.5